Another attempt to fix fontstash.h warnings

+++ modules/core/src/main/include/nanovg/fontstash.h
@@ -863,7 +863,7 @@ int fonsAddFont(FONScontext* stash, const char* name, const char* path)
 	fseek(fp,0,SEEK_SET);
 	data = (unsigned char*)malloc(dataSize);
 	if (data == NULL) goto error;
-	if ( fread(data, 1, dataSize, fp) != dataSize ) goto error;
+	if ( (int)fread(data, 1, dataSize, fp) != dataSize ) goto error;
 	fclose(fp);
 	fp = 0;
